Refactor JS App
